refactor entrypoint
