--- codegen.ts
@@ -1,4 +1,0 @@
-import schemaString from './src/typeDefs';
-import { buildSchema } from 'graphql';
-
-export default buildSchema(schemaString);
