+++ src/index.ts
@@ -1,29 +1,13 @@
 import './env';
-import { ApolloServer, AuthenticationError, ValidationError } from 'apollo-server-express';
-import typeDefs from './typeDefs';
-import resolvers from './resolvers';
 import express from 'express';
-import { makeExecutableSchema } from '@graphql-tools/schema';
 import getDbConnection from './db';
-import { CollectionAccessFunction, Languages, ResolverContextBase } from './types/graphql';
-import languageParser from 'accept-language-parser';
 import bodyParser from 'body-parser';
 import router from './router';
 import { ApiError } from './errorTypes';
 import errorHandler from './middlewares/errorHandler';
-import multilingualDirective from './directives/multilingual';
-import paginationDirective from './directives/pagination';
-import fromFieldDirective from './directives/fromField';
-import authCheckDirective from './directives/authСheck';
-import adminCheckDirective from './directives/adminCheck';
-import dataLoaderDirective from './directives/dataloaders';
 import * as Sentry from '@sentry/node';
-import { GraphQLError } from 'graphql';
-import jwtHelper from './utils/jwt';
 import DataLoaders from './dataLoaders';
-import globalIdResolver from './globalIdResolver';
-import toGlobalIdDirective from './directives/toGlobalId';
-import createDirectiveDefault from '@codexteam/graphql-directive-default';
+import Server from './server';
 
 console.log('⚡️ Server starting');
 
@@ -58,64 +42,12 @@ Sentry.init({ dsn: process.env.SENTRY_DSN });
    */
   app.use(router);
 
-  const schema = makeExecutableSchema({
-    typeDefs,
-    resolvers,
-    schemaTransforms: [
-      globalIdResolver,
-      toGlobalIdDirective('toGlobalId'),
-      paginationDirective('pagination'),
-      multilingualDirective('multilingual'),
-      fromFieldDirective('fromField'),
-      authCheckDirective('authCheck'),
-      adminCheckDirective('adminCheck'),
-      dataLoaderDirective('dataLoader'),
-      createDirectiveDefault().schemaTransformer,
-    ],
-  });
+  /**
+   * Setup GraphQL server
+   */
+  const server = new Server(dbConnection, dataLoaders);
 
-  const apolloServer = new ApolloServer({
-    schema,
-    formatError: (error: GraphQLError): GraphQLError => {
-      if (error.originalError instanceof ApiError && error.extensions) {
-        error.extensions.code = error.originalError.code;
-      }
-
-      const errorsWhitelist = [ValidationError, AuthenticationError, ApiError];
-
-      const isCaptureNeeded = !errorsWhitelist.some(ErrorType => error.originalError instanceof ErrorType);
-
-      if (isCaptureNeeded) {
-        Sentry.captureException(error);
-      }
-
-      return error;
-    },
-    playground: true,
-    async context({ req }): Promise<ResolverContextBase> {
-      let languages = [ Languages.RU ];
-
-      const languageHeader = req.headers['accept-language'];
-
-      if (languageHeader) {
-        languages = languageParser.parse(languageHeader ? languageHeader.toString() : '').map((language) => {
-          return language.code.toUpperCase() as Languages;
-        });
-      }
-
-      const tokenData = jwtHelper.getDataFromHeader(req.headers.authorization);
-
-      return {
-        db: dbConnection,
-        languages,
-        tokenData,
-        dataLoaders,
-        collection: (name => dbConnection.collection(name)) as CollectionAccessFunction,
-      };
-    },
-  });
-
-  apolloServer.applyMiddleware({ app });
+  server.setupMiddleware(app);
 
   /**
    * Setup sentry error handler
@@ -132,6 +64,6 @@ Sentry.init({ dsn: process.env.SENTRY_DSN });
   app.use(errorHandler);
 
   app.listen({ port: process.env.PORT }, () =>
-    console.log(`🚀 Server ready at http://localhost:${process.env.PORT}${apolloServer.graphqlPath}`)
+    console.log(`🚀 Server ready at http://localhost:${process.env.PORT}${server.apollo.graphqlPath}`)
   );
 })().catch(e => console.error('Error during server starting: ', e));
